Refactor: 암호화 수정, 추가 수정 필요
