+++ bin/main/mapper/user.xml
@@ -25,4 +25,8 @@
    <select id="findByUsernameAndPassword" resultType="shop.mtcoding.job.model.user.User">
       select * from user_tb where username = #{username} and password = #{password}         
    </select>
+
+   <select id="findSaltById" resultType="shop.mtcoding.job.model.user.User">
+      select salt = #{salt} from user_tb where id = #{id}  
+   </select>
 </mapper>
