+++ src/main/java/shop/mtcoding/job/JobApplication.java
@@ -2,8 +2,9 @@ package shop.mtcoding.job;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
 
-@SpringBootApplication
+@SpringBootApplication(exclude = SecurityAutoConfiguration.class)
 public class JobApplication {
 
 	public static void main(String[] args) {
