+++ src/main/java/shop/mtcoding/job/dto/user/UserReqDto.java
@@ -9,7 +9,10 @@ public class UserReqDto {
     public static class JoinReqDto {
         private String username;
         private String password;
+        private String name;
         private String email;
+        private String contact;
+        private String profile;
     }
 
     @Setter
