+++ src/main/resources/db/data.sql
@@ -1,4 +1,4 @@
-insert into user_tb(username, password, name, email, contact, created_at) values('ssar', '03ac674216f3e15c761ee1a5e255f067953623c8b388b4459e13f978d7c846f4', '동식', 'ssar@nate.com', '01011112222', now());
-insert into user_tb(username, password, name, email, contact, created_at) values('cos', '03ac674216f3e15c761ee1a5e255f067953623c8b388b4459e13f978d7c846f4', '영식', 'cos@nate.com', '01022223333', now());
+insert into user_tb(username, password, name, email, contact, created_at) values('ssar', '03ac674216f3e15c761ee1a5e255f067953623c8b388b4459e13f978d7c846f4_{bcrypt}$2a$10$XUJ15pW2ACIGXZzOeZd91.uLrn6G3uWr/K9YK7tH28m8mXtV6wNWa', '동식', 'ssar@nate.com', '01011112222', now());
+insert into user_tb(username, password, name, email, contact, created_at) values('cos', '03ac674216f3e15c761ee1a5e255f067953623c8b388b4459e13f978d7c846f4_{bcrypt}$2a$10$XUJ15pW2ACIGXZzOeZd91.uLrn6G3uWr/K9YK7tH28m8mXtV6wNWa', '영식', 'cos@nate.com', '01022223333', now());
 
 commit;
