+++ src/test/java/shop/mtcoding/job/PasswordEncoderTest.java
@@ -1,13 +1,16 @@
-package shop.mtcoding.job.util;
+package shop.mtcoding.job;
 
+import org.junit.jupiter.api.Test;
 import org.springframework.security.crypto.factory.PasswordEncoderFactories;
 import org.springframework.security.crypto.password.PasswordEncoder;
 
-public class saltEncoder {
-    public static String salt() {
+public class PasswordEncoderTest {
+    @Test
+    public void salt_test() {
         String input = "5thProject";
         PasswordEncoder passwordEncoder = PasswordEncoderFactories.createDelegatingPasswordEncoder();
         String saltPassword = passwordEncoder.encode(input);
-        return saltPassword;
+        System.out.println("테스트 : " + saltPassword);
+
     }
 }
